cleanup + popcount at depth 1
